refactor: move changelog generator in separate package
