+++ pkg/semrel/semrel_test.go
@@ -2,7 +2,6 @@ package semrel
 
 import (
 	"fmt"
-	"strings"
 	"testing"
 
 	"github.com/Masterminds/semver"
@@ -82,26 +81,5 @@ func TestApplyChange(t *testing.T) {
 				}
 			}
 		})
-	}
-}
-
-func TestGetChangelog(t *testing.T) {
-	commits := []*Commit{
-		{},
-		{SHA: "123456789", Type: "feat", Scope: "app", Message: "commit message"},
-		{SHA: "abcd", Type: "fix", Scope: "", Message: "commit message"},
-		{SHA: "12345678", Type: "yolo", Scope: "swag", Message: "commit message"},
-		{SHA: "12345678", Type: "chore", Scope: "", Message: "commit message", Raw: []string{"", "BREAKING CHANGE: test"}, Change: Change{Major: true}},
-		{SHA: "stop", Type: "chore", Scope: "", Message: "not included"},
-	}
-	latestRelease := &Release{SHA: "stop"}
-	newVersion, _ := semver.NewVersion("2.0.0")
-	changelog := GetChangelog(commits, latestRelease, newVersion)
-	if !strings.Contains(changelog, "* **app:** commit message (12345678)") ||
-		!strings.Contains(changelog, "* commit message (abcd)") ||
-		!strings.Contains(changelog, "#### yolo") ||
-		!strings.Contains(changelog, "```BREAKING CHANGE: test\n```") ||
-		strings.Contains(changelog, "not included") {
-		t.Fail()
 	}
 }
